+++ src/app/page.tsx
@@ -13,7 +13,6 @@ export default async function Page() {
 
   return (
     <LandingPageContainer
-      image="home"
       title="Remnant 2 Toolkit"
       description={
         <>